requirements.txt
gmusicapi
spotipy
num2words
python-dotenv
